FastAPI_Backend/requirements.txt
absl-py==1.0.0

aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asgiref==3.5.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
binaryornot==0.4.4
bleach==4.1.0

cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.3
cryptography==36.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
fastapi==0.75.0
flatbuffers==2.0
fonttools==4.28.5
frozenlist==1.3.0
fsspec==2022.2.0
gast==0.5.3
gcsfs==2022.2.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.7.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.34.2
google-cloud-bigquery-storage==2.13.0
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.6.0
idna==3.3
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==0.10.0
json5==0.9.6
jsonschema==4.3.3

keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.22.0
oauthlib==3.2.0
openai==0.11.5
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandas-gbq==0.14.0
pandas-stubs==1.2.0.40
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
poyo==0.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
proto-plus==1.20.3
protobuf==3.19.4
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pydata-google-auth==1.3.0
Pygments==2.11.2
PyJWT==1.7.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.3
python-dotenv==0.19.2
python-slugify==5.0.2
pytz==2021.3

pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3fs==2022.2.0
s3transfer==0.5.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0

tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
wrapt==1.13.3
yarl==1.7.2
